Initial commit: copy of the original X-CUBE-DSPDEMO
This is a copy of the original X-CUBE-DSPDEMO you can find here:
https://www.st.com/en/embedded-software/x-cube-dspdemo.html

Was copied only what is needed for the STM32F407VET CPU, the most
closest project is Projects/STM32F429I-Discovery.

BSP-wise original discovery and chineese "blackboard" are different,
so only BSP/Components/ili9341 driver was taken which will be changed
in the following patch.

"Middlewares" folder was renamed to "Middleware" because stupid
STM32CubeMX has a bug which deletes the folder.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>